Haciendo remove_quotes3
