+++ Navarro/lectur.c
@@ -6,24 +6,42 @@
 /*   By: joanavar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 18:54:07 by joanavar          #+#    #+#             */
-/*   Updated: 2024/11/12 18:57:22 by joanavar         ###   ########.fr       */
+/*   Updated: 2024/11/19 18:21:16 by joanavar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-
-int	string_type(t_token *token)
+static void    is_caracter_token(char c, t_token **stack)
 {
-	if (token->type == 1)
-		return (1);
-	else if (token->type == 2)
-		return (1);
-	else if (token->type == 3)
-		return (1);
-	else 
-		return (0);
+    char *token;
+
+    token = malloc(sizeof(char *) * 2);
+	if (!token)
+		return ;
+    token[0] = c;
+    token[1] = '\0';
+    get_token(token, stack);
 }
+
+static void    is_redireccion(char *str, int i, t_token **stack)
+{
+    char *token;
+    int j;
+
+    j = 0;
+    token = malloc(sizeof(char *) * 3);
+    while (str[i] == '<' || str[i] == '>') // proteger si te mandan >< o <>;
+    {
+        token[j] = str[i];
+        j++;
+        i++;
+    }
+    token[j] = '\0';
+
+    get_token(token, stack);
+}
+
 
 /*void	remove_quotes(t_token *stack)
 {
@@ -45,27 +63,6 @@ int	string_type(t_token *token)
 	}
 }*/
 
-void	remove_quotes(t_token *stack)
-{
-	t_token *tmp;
-	char	*tmp_content;
-
-	while(stack)
-	{
-		if (string_type(stack) && string_type(stack->next))
-		{
-			tmp = stack->next; // guardo referencia 2do nodo
-			tmp_content = stack->content; // guardo referencia 1ra string
-			stack->content = ft_strjoin(stack->content, stack->next->content); // fusiono 1era y 2da string, pierdoo referencia 1ra
-			stack->next = tmp->next; //asigno siguiente nodo al 3r
-			free(tmp_content); // libero 1ra string simple
-			free(tmp->content); // libero 2da string simple
-			free(tmp); // libero segundo nodo
-		}
-		else
-			stack = stack->next;
-	}
-}
 
 void	lectur_imput(char *str)
 {
@@ -101,20 +98,7 @@ void	lectur_imput(char *str)
 	//remove_quotes(*stack);
 }
 
-void	stack_token(t_token *token, t_token **stack)
-{
-	t_token *last_token;
 
-	if (!(*stack))
-	{
-		*stack = token;
-		token->next = NULL;
-		return ;
-	}
-	last_token = find_last(*stack);
-	last_token->next = token;
-	token->prev = last_token;
-}
 
 int main(int argc, char **argv)
 {
